Manage package,file,service client,server
The module was half functional when I picked it back up.  With this
change, the module may be used with a simple:

    include mcollective

And defaults to installing and managing the latest available packages in
the repository, managing the configuration file, and starting the
mcollective service.

This has been developed and tested on CentOS 5.6 and there is no cross
platform support as of now.

The configuration file is being taken as a top level class parameter
allowing the end user to provide a complete configuration file as a
string variable when declaring the class.  This is a convention I've
been making more and more use of but has much room for improvement.