(#8) Add anchor resources and relationships
This change introduces the Anchor pattern to the MCollective module.
The motivation is to ensure the end user of the module may specify
relationships before and requiring the main mcollective class.

For example, without this change, the following code would not work as
the end user expects:

    node default {
      notify { alpha: }
      ->
      class { mcollective: }
      ->
      notify { omega: }

This change introduces a requirement of the puppetlabs-stdlib module at
version 0.1.6 or later.  This is where the anchor type lives.