maint - avoid confusion, don't supply a client.cfg with securityprovider == ssl
When using the ssl securityprovider each user should have their own
credentials configured in ~/.mcollective set up with the mcollective::user
define or manually.

A shared system client.cfg isn't complete or meaningful in this
context, so this patch ensures the client.cfg is absent when the
securityprovider is ssl.