add last_run to the list of dynamic facts that are filtered out
The fact `last_run` comes from the `puppi` module which is fairly
common.  This patch also adds some simple tests for dynamic fact
filtering.  In the future, the list of fact filters is probably going to
need to be configurable via the API instead of hard coded into an ERB
template.