Replace facts.yaml pattern with cron job
Because there is no way to iterate over all possible mutable facts, and
it's better to refresh more frequently than Puppet may run.
Besides which, dynamic information which SHOULDN'T change every Puppet
run (like uptime) SHOULD still be available to use in filters. It's
difficult/impossible to achieve that with a file/template based
Puppet-generated yaml file.

Note that this was tested against PE, but written in a way that should
be agnostic to PE/POSS.