Exclude `path` fact from facts.yaml
The `path` fact is a somewhat dynamic, having different values depending
on who initiates a puppet agent run (a user or Cron/daemon). This leads
to Mcollective restarting when the fact value changes which is
unnecessary and noisy. Preventing `path` from ending up in facts.yaml
quiets things down. Fixes #129.