Expand docs on plugin installation from source.
The current docs on how to install a plugin from source were a bit short
on examples. This adds a complete example of how to sync an agent plugin
using the mcollective::plugin defined type as well as the directory
structure that needs to be created for it to end up in the right place.