Update full-stack example profile module
Previously a "site" stub module was standing in for an example
full-stack profile module. The site stub was not complete and contained
several errors and non-ideal design aspects. This commit replaces the
example module with one called mco_profile, which is more complete and
has been smoke tested for the server profile and activemq middleware
profile.