(MODULES-773) Use string for host_iteration titles in future parser
In the future parser, integers are not allowed as titles. The range()
function in stdlib explicitly returns integers when the arguments are
digits (even String digits), so they must be cast from integers to
strings to use as titles. String + Integer math is still valid in the
future parser, so this should still work as intended.

It already has unit tests showing this behavior with the standard
parser. rspec-puppet does not yet work with the future parser due to
its use of `import`.