common/config: Set mode on site_libdir.
Currently we're not setting a mode in $site_libdir which basically means
we get whatever Puppet decides it should be.

On our systems this suddenly turned $site_libdir to '0770' meaning users
could not longer read/load/initialise plugins that were installed.

By explicitly setting it to '0644' we get the desired behaviour,
everyone can walk the path and read the files but only root can update
them.