Remove environment parameter from cron resource
Crontab environment variables are global and we do not as a matter of
best practice want to leak environment variables into other cron jobs.
We still want to ensure the correct ruby is used to instantiate the
refresh-mcollective-metadata script but we can accomplish this by
templatizing the shebang line rather than relying on environment
variables in the crontab.

Closes #207