Add back in a PATH specifier to the cron job
There is concern that cron jobs run with a reduced PATH, so we still
want to invoke this script with at least as full a path as the Puppet
service has itself. While we want to avoid the environment parameter to
set it as that is global, we can prefix the command itself in the cron
job with the value of the PATH environment variable to use.