Remove Yaml class serialization information
`puppet facts --render-as yaml` returns a yaml file which is the
serialized version of the object:Puppet::Node::Facts.

Mcollective parses this yaml and tries to make a hash out of it.
We get an error:

  `ERROR -- : yaml_facts.rb:31:in `rescue in block in
  load_facts_from_source' Failed to load yaml facts from
  /etc/puppetlabs/mcollective/facts.yaml: TypeError: no implicit
  conversion of Puppet::Node::Facts into HasH`

So we are remove that class information out of the yaml to be able to
serialize it as a ruby hash.